photographyService/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
tzdata==2022.2
urllib3==1.26.12
